+++ synthesize.mjs
@@ -1,17 +1,9 @@
 import fs from 'fs';
 import path from 'path';
-import { T_IDS } from './ids.mjs';
-import { program } from 'commander';
-import { fnr, dnr, hnr, tnr, dnrAndHnr } from './validator.mjs';
+import {T_IDS} from './ids.mjs';
+import {fnr, dnr, hnr, tnr, dnrAndHnr} from './validator.mjs';
+import {generateRandomSwedishPersonalNumber, isValidSwedishPersonalNumber} from "./swedish.mjs";
 
-program
-    .option('-r, --replace-invalid', 'Replace 11-digit numbers that are not valid with synthetic ones')
-    .option('-d, --dir <path>', 'Directory to process')
-
-program.parse(process.argv);
-const options = program.opts();
-
-const baseDirectory = options.directory ? path.resolve(options.directory) : process.cwd();
 const excludeDirs = ['.idea', '.git', 'target', 'node_modules'];
 const replacementNumbers = T_IDS;
 let usedReplacements = {}; // Object to keep track of used replacement numbers
@@ -20,6 +12,7 @@ function isNumberInvalid(number) {
     // Check if number is invalid in all validation functions
     return ![fnr, dnr, hnr, tnr, dnrAndHnr].some(func => func(number).status === 'valid');
 }
+
 function isValidNNID(number) {
     return fnr(number).status === 'valid';
 }
@@ -47,13 +40,20 @@ function getReplacementNumber(originalNumber) {
     return usedReplacements[originalNumber];
 }
 
-export default function synthesize(dirPath = baseDirectory) {
+function getSwedishReplacementNumber(originalNumber) {
+    if (!usedReplacements[originalNumber]) {
+        usedReplacements[originalNumber] = generateRandomSwedishPersonalNumber();
+    }
+    return usedReplacements[originalNumber];
+}
+
+export default function synthesize({dirPath, swedish, replaceInvalid}) {
     const numberMap = {}; // Map of original numbers to their replacements
     const filesToProcess = [];
-    const validate = options.replaceInvalid ? isNumberInvalid : isValidNNID;
+    const validate = swedish ? isValidSwedishPersonalNumber : replaceInvalid ? isNumberInvalid : isValidNNID;
 
     function traverseFileSystem(currentPath) {
-        const entries = fs.readdirSync(currentPath, { withFileTypes: true });
+        const entries = fs.readdirSync(currentPath, {withFileTypes: true});
 
         for (const entry of entries) {
             const fullPath = path.join(currentPath, entry.name);
@@ -63,12 +63,12 @@ export default function synthesize(dirPath = baseDirectory) {
             } else if (entry.isFile()) {
                 const data = fs.readFileSync(fullPath, 'utf8');
                 let match;
-                const regex = /\b\d{11}\b/g;
+                const regex = swedish ? /\b\d{12}\b/g : /\b\d{11}\b/g;
 
                 while ((match = regex.exec(data))) {
                     const number = match[0];
                     if (validate(number)) {
-                        numberMap[number] = getReplacementNumber(number);
+                        numberMap[number] = swedish ? getSwedishReplacementNumber(number) : getReplacementNumber(number);
                         if (!filesToProcess.includes(fullPath)) {
                             filesToProcess.push(fullPath);
                         }