+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2023-11-13
+### Added
+- Replace valid Swedish personal numbers with --swedish option.
+- Generate random synthetic swedish personal number with --fake-swedish option.
+
 ## [1.1.1] - 2023-11-13
 ### Added
 - Command line options -r and -d.